+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
 ## Fixed
 - Fixed a few issues with profile transfer.
-- Fixed Kui text-only RP nameplates not updating when level text and health were both disabled.
+- Fixed Kui name-only RP nameplates not updating when level text and health were both disabled.
 - Fixed inability to set a pet profile when the pet is first summoned.
 - Fixed TRP tooltip overlapping with the main tooltip if the profile has been caught by the mature filter.
 - Fixed TRP chat links breaking from emote/OOC/speech detection.